Combine cloud/network collectors
